feat: adding bottom sheet
